+++ resources/scripts/backup.sh
@@ -5,10 +5,9 @@ date
 export GITLAB_HOME=/srv/gitlab
 cd $GITLAB_HOME
 docker exec gitlab gitlab-ctl stop
-while docker exec gitlab gitlab-ctl status 2>/dev/null | grep -q "run:"; do
-    sleep 5
-done
 docker-compose down
+docker rm -f gitlab 2>/dev/null
+
 vol_arn="${vol_arn}"
 iam_role_arn="${backup_role_arn}"
 vault_name="${vault_name}"